machine-learning-shroyash-student-marks-prediction/student_marks_prediction/requirements.txt
django>=4.2
djangorestframework>=3.14
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
pillow>=10.0